Wow really improved DRY Connor
