+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".LoginActivity"
+        <activity android:name=".Activities.LoginActivity"
                   android:windowSoftInputMode="stateHidden">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -17,11 +17,11 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity"/>
-        <activity android:name=".ScheduleActivity"/>
-        <activity android:name=".StatisticsActivity"/>
-        <activity android:name=".TaggingActivity"/>
-        <activity android:name=".SignUpActivity"
+        <activity android:name=".Activities.MainActivity"/>
+        <activity android:name=".Activities.ScheduleActivity"/>
+        <activity android:name=".Activities.StatisticsActivity"/>
+        <activity android:name=".Activities.TaggingActivity"/>
+        <activity android:name=".Activities.SignUpActivity"
                   android:windowSoftInputMode="stateHidden"/>
     </application>
