+++ app/src/main/java/pitcherseye/pitcherseye/Activities/MainActivity.java
@@ -1,4 +1,4 @@
-package pitcherseye.pitcherseye;
+package pitcherseye.pitcherseye.Activities;
 
 import android.content.Context;
 import android.content.Intent;
@@ -8,6 +8,8 @@ import android.view.View;
 import android.widget.Button;
 
 import com.google.firebase.auth.FirebaseAuth;
+
+import pitcherseye.pitcherseye.R;
 
 public class MainActivity extends AppCompatActivity {
