+++ app/src/main/java/pitcherseye/pitcherseye/Activities/ScheduleActivity.java
@@ -1,9 +1,11 @@
-package pitcherseye.pitcherseye;
+package pitcherseye.pitcherseye.Activities;
 
 import android.content.Context;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import pitcherseye.pitcherseye.R;
 
 public class ScheduleActivity extends AppCompatActivity {
