+++ app/src/main/java/pitcherseye/pitcherseye/Activities/StatisticsActivity.java
@@ -1,28 +1,22 @@
-package pitcherseye.pitcherseye;
+package pitcherseye.pitcherseye.Activities;
 
 import android.content.Context;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.widget.Button;
 
-public class TaggingActivity extends AppCompatActivity {
+import pitcherseye.pitcherseye.R;
 
-    // Buttons
-
-
-    // Request Code
-    int REQUEST_CODE_CALCULATE = 0;
+public class StatisticsActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_tagging);
+        setContentView(R.layout.activity_statistics);
     }
 
     public static Intent newIntent(Context packageContext) {
-        Intent i = new Intent(packageContext, TaggingActivity.class);
+        Intent i = new Intent(packageContext, StatisticsActivity.class);
         return i;
     }
-
 }
