+++ app/src/main/java/pitcherseye/pitcherseye/Activities/TaggingActivity.java
@@ -1,20 +1,29 @@
-package pitcherseye.pitcherseye;
+package pitcherseye.pitcherseye.Activities;
 
 import android.content.Context;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class StatisticsActivity extends AppCompatActivity {
+import pitcherseye.pitcherseye.R;
+
+public class TaggingActivity extends AppCompatActivity {
+
+    // Buttons
+
+
+    // Request Code
+    int REQUEST_CODE_CALCULATE = 0;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_statistics);
+        setContentView(R.layout.activity_tagging);
     }
 
     public static Intent newIntent(Context packageContext) {
-        Intent i = new Intent(packageContext, StatisticsActivity.class);
+        Intent i = new Intent(packageContext, TaggingActivity.class);
         return i;
     }
+
 }
