+++ app/src/main/res/layout/activity_statistics.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="pitcherseye.pitcherseye.StatisticsActivity">
+    tools:context="pitcherseye.pitcherseye.Activities.StatisticsActivity">
 
     <TextView
         android:id="@+id/Stats"
