+++ app/src/main/res/layout/activity_tagging.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="pitcherseye.pitcherseye.TaggingActivity">
+    tools:context="pitcherseye.pitcherseye.Activities.TaggingActivity">
 
     <TextView
         android:id="@+id/stuff"
